# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management System API's",
        "description": "How to login:\n *   Login using the `/authentication/login` endpoint with your email and password.\n *   Copy the token from the response, specifically from `data.token.value`.\n *   Click the Authorize button in the top right corner.\n *   Enter the token in the format: `Bearer <access_token>`.\n *   Click the Authorize button.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/",
            "description": "LIVE server"
        }
    ],
    "paths": {
        "/authentication/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to systems and return the token and inits data.",
                "description": "Do Login",
                "operationId": "a0b5083c5b1dfa24e85eadffbc31c920",
                "parameters": [
                    {
                        "name": "body",
                        "in": "query",
                        "description": "Login Username & Password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "admin@mail.test"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/authentication/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user.",
                "description": "Do Logout",
                "operationId": "c3f856105a08dfdd9274d4c4ad74339d",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get all roles",
                "description": "Get all roles",
                "operationId": "7b66947f6be16e5ba590107522c43c6d",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "create role",
                "description": "create role",
                "operationId": "bcb0907febbb51b1021efa8e9174ce0d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRole"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "get role by id",
                "description": "get role by id",
                "operationId": "5b38991d928cba24ad388107e8c4a93a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "update role",
                "description": "update role by id",
                "operationId": "cf4bcd0bc1dbd5ec769b7e62a269688a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRole"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "delete role",
                "description": "delete role by id",
                "operationId": "eb1e1d4330bed83f78bb6a7643d4cbc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateRole": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New Role Name"
                    },
                    "permission_ids": {
                        "type": "integer",
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateRole": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New Role Update"
                    },
                    "permission_ids": {
                        "type": "integer",
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Roles",
            "description": "Roles"
        }
    ]
}